fix 'move' title, when user clears cache
