+++ app/src/main/java/org/stepic/droid/events/loading/StartLoadEvent.kt
@@ -1,3 +1,3 @@
 package org.stepic.droid.events.loading
 
-class StartLoadEvent
+class StartLoadEvent(val isMove : Boolean = false)
