+++ app/src/main/java/org/stepic/droid/view/dialogs/LoadingProgressDialogFragment.kt
@@ -1,19 +1,18 @@
-package org.stepic.droid.view.custom
+package org.stepic.droid.view.dialogs
 
 import android.app.Dialog
 import android.os.Bundle
 import android.support.v4.app.DialogFragment
-import org.stepic.droid.R
 
-class MovingProgressDialogFragment : DialogFragment() {
+class LoadingProgressDialogFragment : DialogFragment() {
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         isCancelable = false
-        return LoadingProgressDialog(context, R.string.moving)
+        return LoadingProgressDialog(context)
     }
 
     companion object {
         fun newInstance(): DialogFragment {
-            val fragment = MovingProgressDialogFragment()
+            val fragment = LoadingProgressDialogFragment()
             return fragment
         }
     }
